--- SAPex/Repositories/Google/IGoogleRepositories/IGoogleCalendarEventRepository.cs
@@ -1,10 +1,0 @@
-﻿using System;
-using SAPex.Models;
-using SAPex.Repositories.Google.BaseRepositories;
-
-namespace SAPex.Repositories.Google.IGoogleRepositories
-{
-    public interface IGoogleCalendarEventRepository : GoogleBaseRepository<GoogleCalendarEvent,string>
-    {
-    }
-}
