--- SAPex/Repositories/Google/IGoogleRepositories/IGoogleCalendarRepository.cs
@@ -1,9 +1,0 @@
-﻿using System;
-using SAPex.Repositories.Google.BaseRepositories;
-
-namespace SAPex.Repositories.Google.IGoogleRepositories
-{
-    public interface IGoogleCalendarRepository : GoogleBaseRepository<IGoogleCalendarEventRepository,string>
-    {
-    }
-}
