--- SAPex/Repositories/Google/IGoogleRepositories/IGoogleUserAccessTokenRepository.cs
@@ -1,11 +1,0 @@
-﻿using System;
-using SAPex.Models;
-using SAPex.Repositories.Google.BaseRepositories;
-
-namespace SAPex.Repositories.Google.IGoogleRepositories
-{
-    public interface IGoogleUserAccessTokenRepository : GoogleBaseRepository<GoogleUserAccessToken,int>
-    {
-        public GoogleUserAccessToken FindByEmail(string email);
-    }
-}
