--- SAPex/Services/Google/IGoogleServices/IGoogleCalendarEventService.cs
@@ -1,10 +1,0 @@
-﻿using System;
-using SAPex.Models;
-using SAPex.Services.Google.BaseServices;
-
-namespace SAPex.Services.Google.IGoogleSevices
-{
-    public interface IGoogleCalendarEventService : GoogleBaseService<GoogleCalendarEvent, string>
-    {
-    }
-}
