--- SAPex/Services/Google/IGoogleServices/IGoogleCalendarService.cs
@@ -1,11 +1,0 @@
-﻿using System;
-using SAPex.Models;
-using SAPex.Services.Google.BaseServices;
-
-namespace SAPex.Services.Google.IGoogleSevices
-{
-    public interface IGoogleCalendarService : GoogleBaseService<GoogleCalendar,string>
-    {
-
-    }
-}
