--- SAPex/Services/Google/IGoogleServices/IGoogleOAuthService.cs
@@ -1,16 +1,0 @@
-﻿using SAPex.Models;
-using SAPex.Models.Google;
-
-namespace SAPex.Services.Google.IGoogleSevices
-{
-    public interface IGoogleOAuthService
-    {
-        public string GetRedirectUrl();
-        public GoogleUser GetUserInfo(string email);
-        public GoogleUserAccessToken Add(string email,string code);
-        public GoogleUserAccessToken Update(string email);
-        public GoogleUserAccessToken GetGoogleUserAccessToken(string email);
-        public bool Delete(string email);
-      
-    }
-}
