+++ SAPex/appsettings.json
@@ -2,20 +2,6 @@
   "ConnectionStrings": {
     "DefaultConnection": "Server=127.0.0.1,1433;Database=SAPExDb;User Id=SA;Password=MyPass@word;"
   },
-  "GoogleStrings": {
-    "key": "AIzaSyAxAw_wO6NwJchMksGz3VK4Z81EMjHT3vE",
-    "client_id": "664599309202-k5fc3rvm70q2gdt5omb1ag84ui6au2s2.apps.googleusercontent.com",
-    "client_secret": "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",
-    "project_id": "exadel-recruiter-project",
-    "auth_uri": "https://accounts.google.com/o/oauth2/auth",
-    "token_uri": "https://oauth2.googleapis.com/token",
-    "token_revoke_uri": "https://oauth2.googleapis.com/revoke",
-    "auth_provider_x509_cert_url": "https://www.googleapis.com/oauth2/v1/certs",
-    "redirect_uris": "https://localhost:5001/authorization/callback",
-    "google_calendar_events_uri": "https://www.googleapis.com/calendar/v3/calendars/primary/events",
-    "google_calendar_uri": "https://www.googleapis.com/calendar/v3/calendars/",
-    "google_calendar_list_uri": "https://www.googleapis.com/calendar/v3/users/me/calendarList"
-  },
   "JwtConfig": {
     "secret": "PDv7DrqznYL6nv7DrqzjnQYO9JxIsWdcjnQYL6nu0f",
     "expirationInMinutes": 1440
